Add unit tests for controllers and refactor code
This commit involves adding unit tests for UserController, BlogController, and StorageService. Code refactoring in the BlogService and BlogController files was done to improve code quality and efficiency. The 'searchBlogs' method which was not used was removed. The renaming of 'BlogServiceTest', placed inside a different package for better organization, was also part of this commit. Changes have been made in the 'pom.xml' file to include 'mockito-core' library which was necessary for creating the tests.