Update security configurations and refactor PasswordValidator class
Adjusted the SecurityConfig class to enable CORS for all origins, eliminating the previous restriction to "localhost:4200". In PasswordValidator, the Service annotation replaced the Component, and the method of reading the common passwords file was updated. Renamed said file and removed the 'optional' attribute from the Lombok dependency in pom.xml.